fix: Change score type from Int to Long

+++ composeApp/src/commonMain/kotlin/com/yet/tetris/ui/screens/game/GameScreen.kt
@@ -183,7 +183,7 @@ private fun GamePlayingContent(
 @Composable
 
 private fun GameStatsRow(
-    score: Int,
+    score: Long,
     lines: Int,
     time: Long
 ) {
