+++ core/data/src/commonMain/kotlin/com/yet/tetris/data/model/GameRecordDto.kt
@@ -5,7 +5,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class GameRecordDto(
     val id: String,
-    val score: Int,
+    val score: Long,
     val linesCleared: Int,
     val difficulty: DifficultyDto,
     val timestamp: Long
